XIVY-13806: Extract Error-Section from Java Bean into own Error-Tab
