+++ integrations/standalone/tests/integration/activites/interface/program.spec.ts
@@ -1,9 +1,8 @@
 import { test } from '@playwright/test';
 import { InscriptionView } from '../../../pageobjects/InscriptionView';
-import { GeneralTest, runTest } from '../../parts';
+import { GeneralTest, ProgramInterfaceErrorTest, ProgramInterfaceStartTest, runTest } from '../../parts';
 import type { CreateProcessResult } from '../../../glsp-protocol';
 import { createProcess } from '../../../glsp-protocol';
-import { ProgramInterfaceStartTest } from '../../parts/program-interface-start';
 
 test.describe('Program', () => {
   let view: InscriptionView;
@@ -27,5 +26,9 @@ test.describe('Program', () => {
 
   test('Java Bean', async () => {
     await runTest(view, ProgramInterfaceStartTest);
+  });
+
+  test('Error', async () => {
+    await runTest(view, ProgramInterfaceErrorTest);
   });
 });
