+++ integrations/standalone/tests/integration/activites/interface/rule.spec.ts
@@ -1,9 +1,8 @@
 import { test } from '@playwright/test';
-import { GeneralTest, runTest } from '../../parts';
+import { GeneralTest, ProgramInterfaceErrorTest, runTest } from '../../parts';
 import { InscriptionView } from '../../../pageobjects/InscriptionView';
 import type { CreateProcessResult } from '../../../glsp-protocol';
 import { createProcess } from '../../../glsp-protocol';
-import { RuleInterfaceStartTest } from '../../parts/program-interface-start';
 
 test.describe('Rule', () => {
   let view: InscriptionView;
@@ -25,7 +24,7 @@ test.describe('Rule', () => {
     await runTest(view, GeneralTest);
   });
 
-  test('Start', async () => {
-    await runTest(view, RuleInterfaceStartTest);
+  test('Error', async () => {
+    await runTest(view, ProgramInterfaceErrorTest);
   });
 });
