+++ packages/editor/src/components/parts/index.ts
@@ -30,5 +30,6 @@ export * from './rest/RestErrorPart';
 export * from './rest/RestRequestPart';
 export * from './program/event/ProgramStartPart';
 export * from './program/activity/ProgramInterfaceStartPart';
+export * from './program/activity/ProgramInterfaceErrorPart';
 export * from './program/intermediate/EventPart';
 export * from './program/configuration/ConfigurationPart';
